feat: :sparkles: Guard roles with headers businessId and userId

+++ src/modules/user-business-role/user-business-role.service.ts
@@ -1,7 +1,11 @@
-import { UserBusinessRole } from 'entities/user-business-role.entity';
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  ConflictException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
+import { UserBusinessRole } from 'entities/user-business-role.entity';
 
 @Injectable()
 export class UserBusinessRolesService {
@@ -33,7 +37,7 @@ export class UserBusinessRolesService {
       relations: ['role'],
     });
 
-    if (!userRol) {
+    if (!userRol || userRol.length === 0) {
       throw new NotFoundException('Not found');
     }
 
@@ -43,19 +47,43 @@ export class UserBusinessRolesService {
   async updateRole(userId: string, businessId: string, roleId: number) {
     const relation = await this.ubrRepository.findOne({
       where: { user_id: userId, business_id: businessId },
+      relations: ['role'],
     });
 
     if (!relation) throw new NotFoundException('Relation not found');
+
+    if (relation.role.name === 'OWNER') {
+      throw new ConflictException('Cannot modify user with OWNER role');
+    }
 
     relation.role_id = roleId;
     return this.ubrRepository.save(relation);
   }
 
   async removeUserFromBusiness(userId: string, businessId: string) {
-    await this.ubrRepository.delete({
-      user_id: userId,
-      business_id: businessId,
+    const relations = await this.ubrRepository.find({
+      where: { user_id: userId, business_id: businessId },
+      relations: ['role'],
     });
-    return { message: 'User removed from business' };
+
+    if (!relations || relations.length === 0) {
+      throw new NotFoundException('Relations not found');
+    }
+
+    const rolesToRemove = relations.filter((r) => r.role.name !== 'OWNER');
+
+    if (rolesToRemove.length === 0) {
+      throw new ConflictException('Cannot remove any role (only OWNER exists)');
+    }
+
+    const removed = rolesToRemove.map((r) =>
+      this.ubrRepository.delete({
+        user_id: r.user_id,
+        business_id: r.business_id,
+        role_id: r.role_id,
+      }),
+    );
+
+    return removed;
   }
 }
